start extracting the core client functionality
