+++ extclient/slack_configuration.go
@@ -1,4 +1,4 @@
-package slackclient
+package extclient
 
 import (
 	"fmt"
