+++ internal/slackclient/client.go
@@ -1,22 +1,18 @@
 package slackclient
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
 	"log"
-	"net/http"
-	"net/url"
 	"os"
 	"path"
 	"strconv"
 	"strings"
 	"time"
 
-	"github.com/rneatherway/gh-slack/internal/httpclient"
+	"github.com/rneatherway/gh-slack/extclient"
 
 	"nhooyr.io/websocket"
 )
@@ -123,8 +119,8 @@ type Cache struct {
 type SlackClient struct {
 	cachePath string
 	team      string
-	auth      *SlackAuth
 	cache     Cache
+	client    *extclient.SlackClient
 	log       *log.Logger
 	tz        *time.Location
 }
@@ -140,7 +136,7 @@ func New(team string, log *log.Logger) (*SlackClient, error) {
 	}
 	cachePath := path.Join(dataHome, "gh-slack")
 
-	auth, err := getSlackAuth(team)
+	client, err := extclient.New(team)
 	if err != nil {
 		return nil, err
 	}
@@ -148,7 +144,7 @@ func New(team string, log *log.Logger) (*SlackClient, error) {
 	c := &SlackClient{
 		cachePath: cachePath,
 		team:      team,
-		auth:      auth,
+		client:    client,
 		log:       log,
 		tz:        time.Now().Location(),
 	}
@@ -165,10 +161,7 @@ func Null(team string) (*SlackClient, error) {
 	}
 
 	return &SlackClient{
-		team: team,
-		auth: &SlackAuth{
-			Token: "null",
-		},
+		team:      team,
 		cachePath: cacheFile.Name(),
 		tz:        time.UTC,
 	}, nil
@@ -184,79 +177,21 @@ func (c *SlackClient) UsernameForMessage(message Message) (string, error) {
 	return "ghost", nil
 }
 
-func (c *SlackClient) API(verb, path string, params map[string]string, body string) ([]byte, error) {
-	u, err := url.Parse(fmt.Sprintf("https://%s.slack.com/api/", c.team))
-	if err != nil {
-		return nil, err
-	}
-	u.Path += path
-	q := u.Query()
-	for p := range params {
-		q.Add(p, params[p])
-	}
-	u.RawQuery = q.Encode()
-
-	messageBytes := []byte(body)
-	if err != nil {
-		return nil, fmt.Errorf("failed to unmarshal message: %w", err)
-	}
-
-	reqBody := bytes.NewReader(messageBytes)
-
-	resBody := []byte{}
-
-	for {
-		req, err := http.NewRequest(verb, u.String(), reqBody)
-		if err != nil {
-			return nil, err
-		}
-		// FIXME: this doesn't seem to break non-POST/non-data requests, but migth
-		// be polluting the headers.
-		req.Header.Set("Content-Type", "application/json; charset=utf-8")
-		req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", c.auth.Token))
-		for key := range c.auth.Cookies {
-			req.AddCookie(&http.Cookie{Name: key, Value: c.auth.Cookies[key]})
-		}
-
-		resp, err := httpclient.Client.Do(req)
-		if err != nil {
-			return nil, err
-		}
-
-		resBody, err = io.ReadAll(resp.Body)
-		if err != nil {
-			return nil, err
-		}
-
-		if resp.StatusCode == 429 {
-			s, err := strconv.Atoi(resp.Header["Retry-After"][0])
-			if err != nil {
-				return nil, err
-			}
-			d := time.Duration(s)
-			c.log.Printf("rate limited, waiting %ds", d)
-			time.Sleep(d * time.Second)
-		} else if resp.StatusCode >= 300 {
-			return nil, fmt.Errorf("status code %d, headers: %q, body: %q", resp.StatusCode, resp.Header, body)
-		} else {
-			break
-		}
-	}
-
-	return resBody, nil
+func (c *SlackClient) API(verb, path string, params map[string]string, body []byte) ([]byte, error) {
+	return c.client.API(verb, path, params, body)
 }
 
 func (c *SlackClient) get(path string, params map[string]string) ([]byte, error) {
-	return c.API("GET", path, params, "{}")
+	return c.API("GET", path, params, []byte("{}"))
 }
 
 func (c *SlackClient) post(path string, params map[string]string, msg *SendMessage) ([]byte, error) {
 	messageBytes, err := json.Marshal(msg)
 	if err != nil {
-		return nil, fmt.Errorf("failed to unmarshal message: %w", err)
+		return nil, fmt.Errorf("failed to marshal message: %w", err)
 	}
 
-	return c.API("POST", path, params, string(messageBytes))
+	return c.API("POST", path, params, messageBytes)
 }
 
 func (c *SlackClient) ChannelInfo(id string) (*Channel, error) {
